GeoXp-package {GeoXp}	R Documentation

Interactive exploratory spatial data analysis

Description

GeoXp is a tool for researchers in spatial statistics, spatial econometrics, geography, ecology etc allowing to link dynamically statistical plots with elementary maps. This coupling consists in the fact that the selection of a zone on the map results in the automatic highlighting of the corresponding points on the statistical graph or reversely, the selection of a portion of the graph results in the automatic highlighting of the corresponding points on the map. GeoXp includes tools from different areas of spatial statistics including geostatistics as well as spatial econometrics and point processes. Besides elementary plots like boxplots, histograms or simple scatterplos, GeoXp also couples with maps Moran scatterplots, variogram cloud, Lorentz Curves,... In order to make the most of the multidimensionality of the data, GeoXp includes some dimension reduction techniques such as PCA.

Details

Package: 	 GeoXp

Type: 	 Package

Version: 	 1.6.1

Date: 	 2012-09-07

License: 	 GPL Vesion 2 or later

In the version 1.5.0, GeoXp has adopted the SpatialClass object proposed by Roger Bivand in sp package. The main advantage of using this structure object is on one hand, a SpatialClass object can contain both spatial coordinates and a data.frame of observed variable and on an other hand, it offers the possibility to make spatial analysis using both packages derived from sp as spdep, gstat and GeoXp.
On the map, the coordinates of sites are represented by using the function coordinates included in sp package, which calculates longitude (for x-axis) and latitude (for y-axis), applied on a Spatial Class Object.
In GeoXp, we can find three main groups of functions:
- functions using only one variable: the interest variable is designed by argument name.var, a character corresponding to a column of the data.frame included in sp.obj, i.e. the Spatial Class object. It can be a numeric variable (histomap(), densitymap(), angleplotmap...) or a factor variable (or character) (barmap(),...).
- functions using both several variables: the variables of interest are designed by argument names.var, a vector of character corresponding to columns of the data.frame included in sp.obj. It can be two numeric variables (dblehistomap, dbledensitymap), one numeric variable and one factor (histobarmap(), polyboxplotmap()), several numeric variables (plot3dmap, pcamap() and clustermap()).
- functions using both a variable and a spatial weight matrix created as a nb or listw object (see package spdep).
In the case where sp.obj is a SpatialPolygonDataFrame, user will have the opportunity to draw the polygons of Spatial unit by using the Draw Saptial contours button in the Tk window. User can also give a spatial polygonal contour as background map with option carte: in this case, a spatial polygonal contour is a matrix of numeric values with 2 columns (x and y coordinates of the vertices of the polygons) where polygons are seperated from each other by 3 rows of NaN. The functions (polylist2list() and spdf2list()) convert some spatial objects (Polylist and SpatialPolygonDataFrame) into matrix as decribed above to draw a background map.
Among options which are common to each function, users have the possibility to give a criteria, vector of boolean of size the number of Spatial units, with TRUE on specific sites. These sites are then represented by a green croice on the map by clicking on preselected sites button on the Tk window.
Moreover, users have the possibility to make bubbles and add some graphs (histogram, barplot or scattermap). The potential variables are included in the data.frame of the SpatialObject. Users can choose a proportional symbol mapping: in function plot, we give value sqrt(var). User can choose if a legend has to appear on the map. He could choose then three values represented by bubbles of corresponding sizes.
Finally, users can choose to represent the graphical with different colors using argument col. In the case of factors (as function barmap), users could choose if a legend with corresponding colors will appear on the map. Users can also modify the representation of selected sites on map with argument pch.
Recent functions barnbmap and histnbmap give the opportunity to analyse spatial weight matrix build using functions included in spdep package.

Author(s)

Christine Thomas-Agnan, Yves Aragon, Anne Ruiz-Gazen, Thibault Laurent, Laurianne Robidou Maintainer: Thibault Laurent <thibault.laurent@univ-tlse1.fr>

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

Roger S.Bivand, Edzer J.Pebesma, Virgilio Gomez-Rubio (2009), Applied Spatial Data Analysis with R, Springer.

[Package GeoXp version 1.6.2 Index]