carte {GeoXp}	R Documentation

Drawing a map

Description

The function carte() draws a map with sites of coordinates (long,lat) and represents sites which have been selected in obs. This function is called by most of the functions of GeoXp (this is not an interactive function).

Usage

carte(long, lat, obs, sp.obj=NULL, criteria=NULL, buble=FALSE, cbuble=NULL, nointer=FALSE,
carte=NULL,nocart=FALSE, label="",cex.lab=NULL, symbol=16, lablong="", 
lablat="", method="", W=NULL,couleurs="blue", classe=NULL, legmap=NULL,
legends=list(FALSE,FALSE),labmod="",axis=FALSE)

Arguments

long
a vector x of size n

lat
a vector y of size n

obs
a boolean vector of size n with TRUE on selected sites

sp.obj
a SpatialPolygonsDataFrame object

criteria
a vector of size n of boolean with TRUE on specific sites (these for non interactive selection)

buble
a boolean with TRUE for drawing bubbles, FALSE otherwise

cbuble
vector of size n with size of each site depending on variable with which bubbles are constructed

nointer
a boolean with TRUE for drawing sites selected by criteria, FALSE otherwise

carte
matrix with 2 columns for drawing spatial polygonal contours : x and y coordinates of the vertices of the polygon

nocart
a boolean with TRUE for drawing spatial contours, FALSE otherwise

label
vector of character of size n with name of each site

cex.lab
character size of label

symbol
a vector of symbol which must be equal to the number of group else all sites are printed in pch[1]

lablong
name of the x-axis that will be printed on the map

lablat
name of the y-axis that will be printed on the map

method
Specification for some graphics such as ‘Neighbourplot1’, ‘Neighbourplot2’, ‘Cluster’, ‘Angleplot’, ‘Variocloud’

W
neighbourhood matrix of size n x n, necessary when method=‘Neighbourplot’

couleurs
Colors of the sites

classe
vector of n whith class of sites

legmap
a list of (numeric,numeric,numeric,numeric,numeric,numeric,character) with the sizes and the corresponding values of the 3 bubbles represented in the legend and the variable name of variable choosed

legends
a list of c(boolean,boolean,c(numeric,numeric),c(numeric,numeric)) with TRUE for drawing legends of bubbles (1st argument) or factors (2nd argument) and the coordinates of the location of the upper left corner of the legend box

labmod
Name of factors

axis
a boolean with True for drawing axes on the map

Details

Options permit to represent sites with different colors or symbols, with different sizes of circle (‘bubble’), using a spatial contour (‘carte’), printing names of sites (‘label’),etc. Possibility to draw a legend giving the sizes of bubbles and colors of levels.

Value

No values, only drawing of a map

Author(s)

Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

[Package GeoXp version 1.6.2 Index]