choix.bubble {GeoXp}	R Documentation

Selection of a variable before plotting bubbles

Description

The function choix.bubble() is used in most of the GeoXp functions to initialize some parameters before plotting bubbles.

Usage

choix.bubble(buble,listvar,listnomvar,legends)

Arguments

buble
boolean

listvar
list of variables

listnomvar
names of variables

legends
parameters of plot

Details

This function is not an interactive function.

Value

A list of parameters which permit to use the function carte.

Author(s)

Laurent T.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

[Package GeoXp version 1.6.2 Index]