choix.couleur {GeoXp}	R Documentation

Selection of colors before plotting map

Description

The function choix.couleur() is used in most of the GeoXp functions to initialize some parameters before plotting map in the case of the use of a factor.

Usage

choix.couleur(graphChoice,listvar=NULL,listnomvar=NULL,
varChoice1=NULL,legends,col,pch,spdf=FALSE)

Arguments

graphChoice
kind of graphic chosen

listvar
list of variables

listnomvar
names of variables

varChoice1
the name of the chosen variable

legends
parameters of plot

col
color of plot

pch
symbols

spdf
a boolean

Details

This function is not an interactive function.

Value

A list of parameters which permit to use the function carte.

Author(s)

Laurent T.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

[Package GeoXp version 1.6.2 Index]