graphique {GeoXp}	R Documentation

Drawing a graphic

Description

The function ‘graphique’ draws a specific graphic. Possibility to draw one of these graphics : Histogram, Barplot, Boxplot, Polyboxplot, Scatterplot, Densityplot, Neighbourplot, Lorentz, Moran, Angleplot, Variocloud, Cluster and PCA (graphics of individuals and variables). This function is called in most of the functions of GeoXp (this is not an interactive function).

Usage

graphique(var1, var2, var3, obs, num, graph = "", couleurs = "",
    symbol = 16, labvar = "", nbcol = 10, alpha1, W,
    Xpoly, Ypoly, F, G, opt1 = 1, opt2 = 1, quantiles = 0,
    labmod = "", direct, inertie, label = 0, kernel, obsq, locmoran = FALSE,
    bin = NULL, cex.lab=1, buble=FALSE, cbuble=NULL, legmap=NULL, 
    legends=list(FALSE,FALSE), xlim, ylim)

Arguments

var1
1st variable of size n

var2
2nd variable of size n

var3
3rd variable of size n (used for variocloudmap)

obs
a vector of boolean of size n with sites selected

num
number of windows which must be activated (3 ou 4)

graph
name of graphic which must be drawn : Histogram, Barplot, Boxplot, Polyboxplot, Scatterplot, Densityplot1, Densityplot2, Neighbourplot, Lorentz, Moran, Quadrant, Angleplot, Variocloud, Cluster, Acp1, Acp2

couleurs
Possibilty to change colors on the graphic

symbol
choice of representation of selected sites

labvar
name(s) of variable(s) studied

nbcol
number of cells if the graphic choosen is the histogram (10 by default)

alpha1
regression smoothing paramater

W
Spatial weight matrix

Xpoly
x-coordinates of the vertices of selected polygon

Ypoly
y-coordinates of the vertices of selected polygon

F
Used for Ginimap

G
Used for Ginimap

opt1
Option for adding a curve (regression line or conditionnal quantile depending on the function)

opt2
Option for drawing robust empirical variogram

quantiles
vector which contains the values of alpha for conditionnal quantile

labmod
names of factor if the graphic choosed is a barplot

direct
Used for PCA

inertie
Used for PCA

label
Name of sites

kernel
Name of the kernel choosed in densitymap

obsq
Used for clustermap and barmap

locmoran
Print local moran for each site

bin
The bins chosen to calculate empirical variogram

cex.lab
character size of label

buble
For Local Moran plot only : a boolean with TRUE for drawing bubbles, FALSE otherwise

cbuble
For Local Moran plot only : vector of size n with size of each site depending on variable with which bubbles are constructed

legmap
For Local Moran plot only : a list of (numeric,numeric,numeric,numeric,numeric,numeric,character) with the sizes and the corresponding values of the 3 bubbles represented in the legend and the variable name of variable choosed

legends
For Local Moran plot only : a list of c(boolean,boolean,c(numeric,numeric),c(numeric,numeric)) with TRUE for drawing legends of bubbles (1st argument) or factors (2nd argument) and the coordinates of the location of the upper left corner of the legend box

xlim
the x limits of the plot

ylim
the y limits of the plot

Details

This function is called by any function which draws a graphic. A lot of options are considered because of the large number of graphics proposed

Value

No values, only drawing of a graphic

Author(s)

Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

[Package GeoXp version 1.6.2 Index]