polylist2list {GeoXp}	R Documentation

Extract from a polylist object the vertices of the polygons

Description

The function polylist2list() extracts the vertices of the polygons from a polylist object, seperating polygons from each other by 3 rows of NaN.

Usage

polylist2list(data)

Arguments

data
A polylist object

Details

The user can then represent the coordinates of sites of a polylist object as background map, using the option carte included in all interactive functions of GeoXp.

Value

It returns a matrix of numeric values with 2 columns (x and y coordinates of the vertices of the polygons) where polygons are seperated from each other by 3 rows of NaN.

Author(s)

T. Laurent

References

Bivand R. et al. (2009), Applied Spatial Data Analysis with R, Springer.

See Also

spdf2list

Examples

data(eire)
eire.contours<-polylist2list(eire.polys.utm)

[Package GeoXp version 1.6.2 Index]