rotation {GeoXp}	R Documentation

Transform coordinates of sites using a rotation

Description

The function rotation() is used to modify coordinates of sites by a rotation with an angle equal to angle. This function is ude in deriftmap

Usage

rotation(coords, angle)

Arguments

coords
matrix nx2 of coordinates

angle
value of angle to use in rotation in degree

Details

Let:

x=(cos(theta),-sin(theta))

y=(sin(theta),cos(theta))

nlecoord=coords x cbind(x,y)

Value

matrix nx2 of new coordinates.

Author(s)

Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

See Also

angleplotmap,variocloudmap

[Package GeoXp version 1.6.2 Index]