selectstat {GeoXp}	R Documentation

Selection of values on a graphic

Description

The function selectstat() is used to select sites on a graphic such as Histogram, Barplot, Boxplot, Polyboxplot, Scatterplot, Densityplot, Neighbourplot, Lorentz, Moran, Angleplot, Variocloud, Cluster, or PCA. Most of the GeoXp functions call selectstat.r.

Usage

selectstat(var1, var2, obs, Xpoly, Ypoly, method,
nbcol, W, F, long, lat)

Arguments

var1
1st variable of size n

var2
2nd variable of size n

obs
a boolean vector of size n : TRUE if a site is already selectioned, FALSE otherwise.

Xpoly
X-coordinates of the vertices of selected polygon

Ypoly
Y-coordinates of the vertices of selected polygon

method
name of the graph wich must be drawn among Histogram, Barplot, Boxplot, Polyboxplot, Densityplot, Neighbourplot, Lorentz, Anglepoint, Variopoint

nbcol
nbcol : number of cells for the histogram (10 by default)

W
A matrix of size n x n (for spatial weight)

F
a vector of numeric

long
a vector of x-axis of size n

lat
a vector of y-axis of size n

Details

This function is not an interactive function.

Value

Return a vector of boolean of size n with TRUE if sites have been selected on a graph

Author(s)

Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

[Package GeoXp version 1.6.2 Index]