slider1 {GeoXp}	R Documentation

Scroll bar

Description

Creates a scroll bar for modifying the value of a parameter.

Usage

slider1(fenetre, refresh.code, names.slide, minima, maxima, resolutions,
starts, no=0)

Arguments

fenetre
number of windows

refresh.code
name of function called after modifying the value of a parameter

names.slide
title for scroll bar

minima
minimum value of parameter

maxima
maximum value of parameter

resolutions
scale

starts
Initial Value

no
number of scroll bar

Details

This function is used in the functions which draw a curve to modify the smoothing parameter. This is not an interactive function

Value

Draws a Tk window.

Author(s)

Thomas-Agnan C., Aragon Y., Ruiz-Gazen A., Laurent T., Robidou L.

References

Thibault Laurent, Anne Ruiz-Gazen, Christine Thomas-Agnan (2012), GeoXp: An R Package for Exploratory Spatial Data Analysis. Journal of Statistical Software, 47(2), 1-23.

See Also

densitymap,dbledensitymap

[Package GeoXp version 1.6.2 Index]